requirements.txt
pyudev>=0.22.0
psutil>=5.9.0
requests>=2.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
